// src/schema/root/enum.ts
export enum GlobalStatusType {
    ACTIVE = 'ACTIVE',
    INACTIVE = 'INACTIVE'
}

export enum GlobalError {
    EXCEPTION = 'EXCEPTION',
    NOT_ALLOWED = 'NOT_ALLOWED',
    ALREADY_EXISTS = 'ALREADY_EXISTS',
    VALIDATION_ERROR = 'VALIDATION_ERROR',
    RECORD_NOT_FOUND = 'RECORD_NOT_FOUND',
    INVALID_FIELD_FORMAT = 'INVALID_FIELD_FORMAT',
    INVALID_INPUT = 'INVALID_INPUT',
    INTERNAL_SERVER_ERROR = 'INTERNAL_SERVER_ERROR',
    ALREADY_DELETED = 'ALREADY_DELETED',
    REQUIRED_FIELDS_MISSING = 'REQUIRED_FIELDS_MISSING',
    PAYMENT_ERROR = 'PAYMENT_ERROR',
    NOT_FOUND = 'NOT_FOUND',
    EMAIL_ALREADY_EXISTS = 'EMAIL_ALREADY_EXISTS',
    WEAK_PASSWORD = 'WEAK_PASSWORD',
    ROLE_NOT_FOUND = 'ROLE_NOT_FOUND',
    COMPANY_NOT_FOUND = 'COMPANY_NOT_FOUND',
}

export enum PaymentStatus {
    PENDING = "PENDING",
    SUCCESS = "SUCCESS",
    FAILURE = "FAILURE",
    REFUNDED = "REFUNDED",
    PENDING_PAYMENT = "PENDING_PAYMENT",
}

export enum OrderType {
    PASS = 'PASS',
    GYM_CLASS = 'GYM_CLASS',
    MEMBERSHIP = 'MEMBERSHIP',
    MEMBERSHIP_RENEW = 'MEMBERSHIP_RENEW',
    MEMBERSHIP_FREEZE = 'MEMBERSHIP_FREEZE',
    STORE = 'STORE',
    PRIVATE_COACH = 'PRIVATE_COACH',
    PRIVATE_COACH_CANCELLATION = 'PRIVATE_COACH_CANCELLATION',
    EVENT = 'EVENT',
    LAUNCH_BOOKING_GYM_CLASS = 'LAUNCH_BOOKING_GYM_CLASS',
    QR_SESSION = 'QR_SESSION',
    COURT = 'COURT',
    FREE_TRIAL = 'FREE_TRIAL',
    PAYMENT = 'PAYMENT',
    POS = 'POS',
    CHANGE_MEMBERSHIP = 'CHANGE_MEMBERSHIP',
    TRANSFER_MEMBERSHIP = 'TRANSFER_MEMBERSHIP',
    TERMINATE_MEMBERSHIP = 'TERMINATE_MEMBERSHIP',
    RELOCATE_MEMBERSHIP = 'RELOCATE_MEMBERSHIP',
}

export enum PaymentProvider {
    CHECKOUT = 'CHECKOUT',
    TAP = 'TAP'
}

export enum GatewayCallbackType {
    PAYMENT = 'PAYMENT',
    SAVED_CARD = 'SAVED_CARD',
}

export enum PrivateCoachStatus {
    ACTIVE = "ACTIVE",
    INACTIVE = "INACTIVE",
    CANCELED = "CANCELED"
}

export enum OrderPaymentStatus {
    INACTIVE = 'INACTIVE',
    SUCCESS = 'SUCCESS',
    PENDING = 'PENDING',
    FAILURE = 'FAILURE',
}